move user to own module
